Extract Helm logic into lib
