+++ build.assets/tooling/lib/helm/template.go
@@ -1,3 +1,5 @@
+package helm
+
 /*
  * Teleport
  * Copyright (C) 2023  Gravitational, Inc.
@@ -15,8 +17,6 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-
-package main
 
 import (
 	"bytes"
